+++ isc-frontend/src/components/immigrant-form/index.js
@@ -1,0 +1,1 @@
+export { ImmigrantForm } from "./immigrant-form";
