+++ isc-frontend/src/components/index.js
@@ -5,3 +5,4 @@ export { Header } from "./header";
 export { ContentWrapper } from "./content-wrapper";
 export { FormPage } from "./form-page";
 export { LoadingIcon } from "./loading-icon";
+export { ImmigrantForm } from "./immigrant-form";
