+++ isc-frontend/src/routes/isc/add-immigrant/add-immigrant.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+import { ImmigrantForm } from "../../../components";
+
+export const AddImmigrant = ({ history }) => (
+  <ImmigrantForm history={history} />
+);
