+++ isc-frontend/src/routes/isc/isc.jsx
@@ -4,7 +4,7 @@ import { Route, Switch } from "react-router-dom";
 import { AppContainer, Sidebar } from "../../components";
 import { AuthContext } from "../../contexts/auth";
 import { useSidebarList } from "../../hooks";
-import { ImmigrantForm } from "./add-immigrant";
+import { AddImmigrant } from "./add-immigrant";
 import { AddIscEmployee } from "./add-isc-employee";
 import { Home } from "./home";
 import { Immigrants } from "./immigrants";
@@ -21,7 +21,7 @@ export const ISC = ({ history }) => {
       <Switch>
         <Route exact path="/isc" component={Home} />
         <Route path="/isc/immigrants" component={Immigrants} />
-        <Route exact path="/isc/add-immigrant" component={ImmigrantForm} />
+        <Route exact path="/isc/add-immigrant" component={AddImmigrant} />
         <Route exact path="/isc/add-isc-employee" component={AddIscEmployee} />
         <Route exact path="/isc/profile" component={Profile} />
       </Switch>
